Split device command handlers by domain
