+++ cmd_device_registry.go
@@ -1,0 +1,64 @@
+package main
+
+var deviceCommands = []command{
+	commandByBool("activate", runActivateCommand),
+	commandByBool("ip", runIPCommand),
+	commandByBool("pcap", runPCAPCommand),
+	commandByBool("ps", runPSCommand),
+	commandByBool("install", runInstallCommand),
+	commandByBool("uninstall", runUninstallCommand),
+	commandByBool("lang", runLangCommand),
+	commandByBool("dproxy", runDproxyCommand),
+	commandByBool("info", runInfoCommand),
+	commandByBool("syslog", runSyslogCommand),
+	commandByBool("ostrace", runOSTraceCommand),
+	commandByBool("screenshot", runScreenshotCommand),
+	commandByBool("resetlocation", runResetLocationCommand),
+	commandByBool("devicename", runDeviceNameCommand),
+	commandByBool("apps", runAppsCommand),
+	commandByBool("date", runDateCommand),
+	commandByBool("diagnostics", runDiagnosticsCommand),
+	commandByBool("pair", runPairCommand),
+	commandByBool("readpair", runReadPairCommand),
+	commandByBool("batteryregistry", runBatteryRegistryCommand),
+	commandByBool("reboot", runRebootCommand),
+	commandByBool("shutdown", runShutdownCommand),
+	commandByBool("diskspace", runDiskspaceCommand),
+	commandByBool("batterycheck", runBatteryCheckCommand),
+	commandByBool("erase", runEraseCommand),
+	commandByBool("rsd", runRSDCommand),
+	commandByBool("mobilegestalt", runMobileGestaltCommand),
+	commandByBool("devicestate", runDeviceStateCommand),
+	commandByBool("wifi", runWifiCommand),
+	commandByBool("prepare", runPrepareCommand),
+	commandByBool("set-wallpaper", runSetWallpaperCommand),
+	commandByBool("get-wallpaper", runGetWallpaperCommand),
+	commandByBool("get-icon-layout", runGetIconLayoutCommand),
+	commandByBool("set-icon-layout", runSetIconLayoutCommand),
+	commandByBool("crash", runCrashCommand),
+	commandByBool("instruments", runInstrumentsCommand),
+	commandByBool("image", runImageCommand),
+	commandByBool("assistivetouch", runAssistiveTouchCommand),
+	commandByBool("voiceover", runVoiceOverCommand),
+	commandByBool("zoom", runZoomCommand),
+	commandByBool("lockdown", runLockdownCommand),
+	commandByBool("setlocation", runSetLocationCommand),
+	commandByBool("setlocationgpx", runSetLocationGPXCommand),
+	commandByBool("timeformat", runTimeFormatCommand),
+	commandByBool("httpproxy", runHTTPProxyCommand),
+	commandByBool("profile", runProfileCommand),
+	commandByBool("forward", runForwardCommand),
+	commandByBool("launch", runLaunchCommand),
+	commandByBool("sysmontap", runSysmontapCommand),
+	commandByBool("memlimitoff", runMemlimitOffCommand),
+	commandByBool("kill", runKillCommand),
+	commandByBool("runtest", runTestCommand),
+	commandByBool("runxctest", runXCTestCommand),
+	commandByBool("runwda", runWDACommand),
+	commandByBool("ax", runAXCommand),
+	commandByBool("resetax", runResetAXCommand),
+	commandByBool("debug", runDebugCommand),
+	commandByBool("file", runFileCommand),
+	commandByBool("fsync", runFsyncCommand),
+	commandByBool("devmode", runDevModeCommand),
+}
